cloud_functions/cloud_datastore/requirements.txt
# Function dependencies, for example:
# package>=version

Flask
flask-cors
redis
google-api-python-client
google-cloud-datastore